:sparkles: feat(contactController,validation): add uuid, email and category_id validations

+++ api/src/app/controllers/contact-controller.js
@@ -1,4 +1,5 @@
 import { contactRepository } from '../repositories/contact-repository.js'
+import { isValidUUID } from '../utils/isValidUUID.js'
 
 class ContactController {
   async index(request, response) {
@@ -12,10 +13,13 @@ class ContactController {
   async show(request, response) {
     const { id } = request.params
 
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id' })
+    }
+
     const contact = await contactRepository.findById(id)
 
     if (!contact) {
-      // 404: Not Found
       return response.status(404).json({ error: 'Contact not found' })
     }
 
@@ -29,19 +33,25 @@ class ContactController {
       return response.status(400).json({ error: 'Name is required' })
     }
 
-    const contactExists = await contactRepository.findByEmail(email)
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category' })
+    }
 
-    if (contactExists) {
-      return response
-        .status(400)
-        .json({ error: 'This e-mail is already in use' })
+    if (email) {
+      const contactExists = await contactRepository.findByEmail(email)
+
+      if (contactExists) {
+        return response
+          .status(400)
+          .json({ error: 'This e-mail is already in use' })
+      }
     }
 
     const contact = await contactRepository.create({
       name,
-      email,
+      email: email || null,
       phone,
-      category_id,
+      category_id: category_id || null,
     })
 
     response.status(201).json(contact)
@@ -51,14 +61,24 @@ class ContactController {
     const { id } = request.params
     const { name, email, phone, category_id } = request.body
 
-    const contactExists = await contactRepository.findById(id)
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id' })
+    }
 
-    if (!contactExists) {
-      return response.status(404).json({ error: 'Contact not found' })
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category' })
     }
 
     if (!name) {
       return response.status(400).json({ error: 'Name is required' })
+    }
+
+    if (email) {
+      const contactExists = await contactRepository.findById(id)
+
+      if (!contactExists) {
+        return response.status(404).json({ error: 'Contact not found' })
+      }
     }
 
     const contactByEmail = await contactRepository.findByEmail(email)
@@ -71,9 +91,9 @@ class ContactController {
 
     const contact = await contactRepository.update(id, {
       name,
-      email,
+      email: email || null,
       phone,
-      category_id,
+      category_id: category_id || null,
     })
 
     response.status(200).json(contact)
@@ -82,9 +102,12 @@ class ContactController {
   async delete(request, response) {
     const { id } = request.params
 
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id' })
+    }
+
     await contactRepository.delete(id)
 
-    // 204: No Content
     response.sendStatus(204)
   }
 }
